moved partner list to PartnerList.json and alphabetized it by partnerId; added JSON schema for it
