+++ .vscode/settings.json
@@ -1,0 +1,66 @@
+{
+   "json.schemas": [
+      {
+         "fileMatch": [
+            "/PartnerList.json"
+         ],
+         "schema": {
+            "$schema": "http://json-schema.org/draft-07/schema#",
+            "definitions": {
+               "Partner": {
+                  "properties": {
+                     "additionalVanityUrls": {
+                        "description": "Optional additional vanity URLs that will resolve to this partner.\nNote that `partnerId` will itself always be a valid URL (`https://powerthepolls.org/{partnerId}`),\nso this is only needed if there are alternative URLs desired.",
+                        "items": {
+                           "type": "string"
+                        },
+                        "type": "array"
+                     },
+                     "customSignupFormField": {
+                        "description": "The label for an additional field to be displayed on the signup form. If not provided, no additional field\nwill be displayed.",
+                        "type": "string"
+                     },
+                     "excludeFromPartnerList": {
+                        "description": "Optional case (typically for certain more partisan partners) that we want to exclude from the public Partner List page\ndespite them having a `logo` assigned.",
+                        "type": "boolean"
+                     },
+                     "isFoundingPartner": {
+                        "description": "If `true` this partner is a Founding Partner",
+                        "type": "boolean"
+                     },
+                     "logo": {
+                        "description": "A logo for partners that should be displayed on the `/partners` page (page-partners.tsx). Images are sourced\nfrom `/site/public/assets/images/partners/*`). If no logo is provided, the partner will no appear on the\nPartners page.",
+                        "type": "string"
+                     },
+                     "logoAppearsOnLandingPage": {
+                        "description": "Whether to include the partner `logo` on the landing page of the partner's vanity URL (either\n`https://powerthepolls.org/{partnerId}` or `https://powerthepolls.org/{vanityUrls}`).",
+                        "type": "boolean"
+                     },
+                     "logoIsDark": {
+                        "description": "If true, the `logo`, when displayed, will receive a white background. If the logo is already on a white background\nthen set this to `true` so that the corners will be rounded.",
+                        "type": "boolean"
+                     },
+                     "name": {
+                        "description": "The full name of this partner.",
+                        "type": "string"
+                     },
+                     "optUserOutOfChase": {
+                        "description": "malachi: it seems tthis is really just turning into a `showPartnerNameInPrivacyDisclaimer` value and that's all",
+                        "type": "boolean"
+                     },
+                     "partnerId": {
+                        "description": "The key to use for thie source in ActionKit and for a vanity URL `https://powerthepolls.org/{partnerId}`\nfor the partner which will include their partnerID when the form is submitted",
+                        "type": "string"
+                     }
+                  },
+                  "type": "object"
+               }
+            },
+            "items": {
+               "$ref": "#/definitions/Partner"
+            },
+            "type": "array"
+         }
+      }
+   ]
+}
