+++ site/src/components/page-faq/readme.md
@@ -7,9 +7,10 @@
 
 ## Properties
 
-| Property | Attribute | Description                                                               | Type                                                   | Default     |
-| -------- | --------- | ------------------------------------------------------------------------- | ------------------------------------------------------ | ----------- |
-| `data`   | --        | A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx | `undefined \| { question: string; answer: string; }[]` | `undefined` |
+| Property     | Attribute    | Description                                                               | Type                                                   | Default     |
+| ------------ | ------------ | ------------------------------------------------------------------------- | ------------------------------------------------------ | ----------- |
+| `data`       | --           | A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx | `undefined \| { question: string; answer: string; }[]` | `undefined` |
+| `page_title` | `page_title` | The page's title                                                          | `string \| undefined`                                  | `undefined` |
 
 
 ----------------------------------------------
