+++ site/src/data/Social.ts
@@ -1,10 +1,5 @@
-export interface SocialInfo {
-   readonly name: string;
-   readonly cta: string;
-   readonly image: string;
-   readonly url: string;
-   readonly eventKey: string;
-}
+import { SocialInfo } from "./types";
+
 /**
  * Social share links for the left nav
  * see: social-share.tsx
