+++ site/tsconfig.json
@@ -7,6 +7,7 @@
       "target": "es2017",
       "module": "esnext",
       "moduleResolution": "node",
+      "resolveJsonModule": true,
       "lib": [
          "dom",
          "dom.iterable",
